Add files via upload
added images in the HTML. In CSS: updated the code for the images, changed the text color, added more text, added code to keep the footer at the bottom of the page. [Next task will be to figure out how to make the link stay on the text portion of the footer and not the footer line along the bottom of the page or the link be just one of the words in the footer text. Also working to add proper hover over the images.]